Crate problem026

Problem 26 (Reciprocal cycles)

Solution detail

The most interesting element is in how to find the cycle length of 1/d:

The digits in the decimal expansion of the fraction 1/d can be found by performing a long division of d into the number 100000.... This leads to the following way of calculating the length of the cycle.

Remove all factors of 2 or 5 from d. This is because dividing by 2 or 5 is equivalent to first multipying by 5 or 2 and subsequently dividing by 10, neither of which changes the cycle length.

With the resulting number d', find the order of 10 modulo d'. This is the length of the period of 1/d', since it is when the remainders of dividing successive powers of 10 by d' start to repeat in sequence. This exactly means that the digits in the computed decimal expansion (which is, after all, found by long division) will repeat.

We can further note that, since there are only d - 1 possible remainders when dividing by d, the length of the cycle of 1/d is at most d - 1. If we start at large values of d and work our way down, we may be able to terminate early when the later values of d will be too small.

The final optimisation is to only consider those values of d which are not divisible by 2 or 5, since other values of d will be equivalent to smaller values of d which result from dividing out the 2 and 5.

Functions

answer
Solve the problem, returning the answer as a String

cycle_length
Find the length of the cycle of the decimal 1/d, where d is coprime to 10. Requires a prime sieve capable of factorising numbers up to d (that is, containing primes up to the square root of d).

main

solve
Find the number below the given limit with the longest repeating decimal.